requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.11.0
fonttools==4.28.2
idna==3.3
kaggle==1.5.12
kiwisolver==1.3.2
matplotlib==3.5.0
numpy==1.21.4
opencv-python==4.5.4.60
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
scipy==1.7.2
seaborn==0.11.2
setuptools-scm==6.3.2
six==1.16.0
text-unidecode==1.3
tomli==1.2.2
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
typing_extensions==4.0.0
urllib3==1.26.7
